+++ tools/Quests/quests.schema.d.ts
@@ -1,0 +1,19 @@
+/* This file was automatically generated. Do not modify it directly. */
+
+/**
+ * A schema for quests parsed from RuneLite's Quest enum.
+ */
+export type Quests = {
+  /**
+   * RuneLite quest ID used by script engine
+   */
+  id: number;
+  /**
+   * Enum constant name in RuneLite's Quest.java
+   */
+  enum: string;
+  /**
+   * Display name of the quest
+   */
+  name: string;
+}[];